+++ db/testserver_sp_power_module.sql
@@ -6,7 +6,7 @@ delimiter $$
 use testserverdb$$
 create procedure sp_add_power_module
 (
-		in p_numero_serie					int
+	in p_numero_serie					bigint(15)
 )
 
 begin
@@ -31,66 +31,66 @@ delimiter $$
 use testserverdb$$
 create procedure sp_add_log_power_module
 (
-        in p_resultado_teste						        varchar(45),
-		in p_numero_serie_power_module       		        int,
-		in p_iload0											double,
-		in p_iload1											double,
-		in p_iload2											double,
-		in p_iload3											double,
-		in p_iload4											double,
-		in p_iload5											double,
-        in p_vload0											double,
-		in p_vload1											double,
-		in p_vload2											double,
-		in p_vload3											double,
-		in p_vload4											double,
-		in p_vload5											double,
-        in p_vdclink0										double,
-		in p_vdclink1										double,
-		in p_vdclink2										double,
-		in p_vdclink3										double,
-		in p_vdclink4										double,
-		in p_vdclink5										double,
-        in p_temperatura0   								double,
-		in p_temperatura1 								    double,
-		in p_temperatura2 								    double,
-		in p_temperatura3 								    double,
-		in p_temperatura4 								    double,
-		in p_temperatura5 								    double,
-		in p_details										text
+    in p_resultado_teste						        varchar(45),
+	in p_numero_serie_power_module       		        bigint(15),
+	in p_iload0											double,
+	in p_iload1											double,
+	in p_iload2											double,
+	in p_iload3											double,
+	in p_iload4											double,
+	in p_iload5											double,
+    in p_vload0											double,
+	in p_vload1											double,
+	in p_vload2											double,
+	in p_vload3											double,
+	in p_vload4											double,
+	in p_vload5											double,
+    in p_vdclink0										double,
+	in p_vdclink1										double,
+	in p_vdclink2										double,
+	in p_vdclink3										double,
+	in p_vdclink4										double,
+	in p_vdclink5										double,
+    in p_temperatura0   								double,
+	in p_temperatura1 								    double,
+	in p_temperatura2 								    double,
+	in p_temperatura3 								    double,
+	in p_temperatura4 								    double,
+	in p_temperatura5 								    double,
+	in p_details										text
 )
 
 begin
 insert into LogModuloPotencia
 (
-		data,
+	data,
     resultado_teste,
     numero_serie_modulo_potencia,
-		iload0,
-		iload1,
-		iload2,
-		iload3,
-		iload4,
-		iload5,
-        vload0,
-		vload1,
-		vload2,
-		vload3,
-		vload4,
-		vload5,
-        vdclink0,
-		vdclink1,
-		vdclink2,
-		vdclink3,
-		vdclink4,
-		vdclink5,
-        temperatura0,
-		temperatura1,
-		temperatura2,
-		temperatura3,
-		temperatura4,
-		temperatura5,
-		details
+	iload0,
+	iload1,
+	iload2,
+	iload3,
+	iload4,
+	iload5,
+    vload0,
+	vload1,
+	vload2,
+	vload3,
+	vload4,
+	vload5,
+    vdclink0,
+	vdclink1,
+	vdclink2,
+	vdclink3,
+	vdclink4,
+	vdclink5,
+    temperatura0,
+	temperatura1,
+	temperatura2,
+	temperatura3,
+	temperatura4,
+	temperatura5,
+	details
 )
 values
 (
